+++ src/App.css
@@ -220,3 +220,11 @@ body {
 .add-to-cart-icon:hover {
   color: red; /* Change to the desired hover color */
 }
+
+.formulario {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+}
