--- src/Checkout.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-function Checkout() {
-  return (
-    <div>
-      <h1>Página de checkout</h1>
-      {/* Contenido de la página de checkout */}
-    </div>
-  );
-}
-
-export default Checkout;
